Move more services to GraphQL context
