+++ src/main/java/org/opentripplanner/standalone/server/DefaultServerRequestContext.java
@@ -12,8 +12,8 @@ import org.opentripplanner.raptor.configure.RaptorConfig;
 import org.opentripplanner.routing.DefaultRoutingService;
 import org.opentripplanner.routing.algorithm.raptoradapter.transit.TransitTuningParameters;
 import org.opentripplanner.routing.algorithm.raptoradapter.transit.TripSchedule;
+import org.opentripplanner.routing.api.RoutingService;
 import org.opentripplanner.routing.api.request.RouteRequest;
-import org.opentripplanner.routing.api.request.RoutingService;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.service.worldenvelope.WorldEnvelopeService;
 import org.opentripplanner.standalone.api.HttpRequestScoped;
